requirements.txt
Flask==3.0.2
gunicorn==21.2.0
APScheduler==3.10.4
pyserial==3.5
apprise==1.9.3
pytest==8.0.2
python-dotenv==1.0.1 